pyproject.toml
[tool.poetry]
name = "psychological-assessments-saas"
version = "1.0.0"
description = "Plataforma SaaS multitenant de assessments psicológicos e gestão de PDI"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.3"
psycopg2-binary = "2.9.9"
redis = "5.0.1"
celery = "5.3.4"
django-allauth = "0.57.0"
django-cors-headers = "4.3.1"
django-csp = "3.7"
django-ratelimit = "4.1.0"
django-storages = "1.14.2"
django-fernet-fields = "0.6"
weasyprint = "61.2"
pandas = "2.1.4"
xlsxwriter = "3.1.9"
pillow = "10.1.0"
python-decouple = "3.8"
gunicorn = "21.2.0"
whitenoise = "6.6.0"
boto3 = "1.34.0"
stripe = "7.8.0"
paypalrestsdk = "1.13.3"
pyjwt = "2.8.0"
cryptography = "41.0.8"
requests = "2.31.0"
python-dateutil = "2.8.2"
pyyaml = "6.0.1"
markdown = "3.5.1"
bleach = "6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-django = "4.7.0"
pytest-cov = "4.1.0"
pytest-mock = "3.12.0"
factory-boy = "3.3.0"
playwright = "1.40.0"
pre-commit = "3.6.0"
ruff = "0.1.7"
mypy = "1.7.1"
django-types = "0.19.1"
types-redis = "4.6.0.11"
types-requests = "2.31.0.20"
types-pyyaml = "6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["migrations", "static"]

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "D", "UP", "B", "A", "S", "T20", "Q"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "S101"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"